Migrated types to enums and added categories
